refactor: mvc패턴을 약간 변형하여 기능별로 모듈 분리
model은 gameLogic.js, inputCheck.js
MissionUtils 모듈로 입출력을 하기 때문에 model이 viewer의 역할도 하도록 mvc를 m+v/c 형식으로 변형
controller는 controller.js
